requirements.txt
argon2-cffi==16.3.0
arrow==0.10.0
awesome-slugify==1.6.5
binaryornot==0.4.4
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
Collectfast==0.5.2
cookiecutter==1.5.1
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.33.0
django-anymail==0.11.1
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.4
django-extensions==1.9.0
django-model-utils==3.0.0
django-redis==4.8.0
django-storages==1.6.5
future==0.16.0
gunicorn==19.7.1
idna==2.6
Jinja2==2.9.6
jinja2-time==0.2.0
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
poyo==0.4.1
psycopg2==2.7.3.1
pycparser==2.18
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
regex==2017.7.28
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.21
urllib3==1.22
whichcraft==0.4.1
whitenoise==3.3.0
